move quantization kernels and wrappers to their own files
