+++ aiter/ops/triton/_triton_kernels/attention/fav3_sage_attention.py
@@ -1461,263 +1461,3 @@ def sage_fwd(
 
     tl.store(o_ptrs, acc.to(Out.dtype.element_ty), mask=o_ptrs_mask)
 
-
-def sage_quant(
-    q,
-    k,
-    v,
-    FP8_TYPE,
-    FP8_MAX,
-    BLKQ=128,
-    BLKK=64,
-    sm_scale=None,
-    layout="bshd",
-    smooth_k=True,
-):
-    """
-    Quantize Q and K tensors to INT8 with per-block scaling.
-
-    Args:
-        q: Query tensor
-        k: Key tensor
-        km: Optional pre-computed K smoothing factors (if None and smooth_k=True, will be computed)
-        BLKQ: Block size for Q quantization
-        BLKK: Block size for K quantization
-        sm_scale: Softmax scale factor (defaults to head_dim^-0.5)
-        layout: Either "bshd" or "bhsd"
-        smooth_k: Whether to apply SageAttention-style smoothing to K tensor (default: True)
-
-    Returns:
-        q_int8: Quantized Q tensor
-        q_scale: Per-block scales for Q
-        k_int8: Quantized K tensor
-        k_scale: Per-block scales for K
-        k_smooth: K smoothing factors applied (or None if smooth_k=False)
-    """
-    q_int8 = torch.empty_like(q, dtype=torch.int8, device=q.device)
-    k_int8 = torch.empty_like(k, dtype=torch.int8, device=k.device)
-    v_fp8 = torch.empty_like(v, dtype=FP8_TYPE, device=v.device)
-
-    if layout == "bhsd":
-        b, h_qo, qo_len, head_dim = q.shape
-        _, h_kv, kv_len, _ = k.shape
-
-        stride_bz_q, stride_h_q, stride_seq_q = q.stride(0), q.stride(1), q.stride(2)
-        stride_bz_k, stride_h_k, stride_seq_k = k.stride(0), k.stride(1), k.stride(2)
-
-    elif layout == "bshd":
-        b, qo_len, h_qo, head_dim = q.shape
-        _, kv_len, h_kv, _ = k.shape
-
-        stride_bz_q, stride_h_q, stride_seq_q = q.stride(0), q.stride(2), q.stride(1)
-        stride_bz_k, stride_h_k, stride_seq_k = k.stride(0), k.stride(2), k.stride(1)
-    else:
-        raise ValueError(f"Unknown tensor layout: {layout}")
-    Q_NUM_BLKS = (qo_len + BLKQ - 1) // BLKQ
-    K_NUM_BLKS = (kv_len + BLKK - 1) // BLKK
-
-    # Apply K tensor smoothing following SageAttention approach
-    if smooth_k:
-        k = k - k.mean(dim=1 if layout == "bshd" else 2, keepdim=True)
-
-    q_scale = torch.empty((b, h_qo, Q_NUM_BLKS), device=q.device, dtype=torch.float32)
-    k_scale = torch.empty((b, h_kv, K_NUM_BLKS), device=q.device, dtype=torch.float32)
-
-    v_scale = v.abs().amax(dim=1 if layout == "bshd" else 2).to(torch.float32) / FP8_MAX
-
-    if sm_scale is None:
-        sm_scale = head_dim**-0.5
-
-    q_task_count = b * h_qo * Q_NUM_BLKS
-    k_task_count = b * h_kv * K_NUM_BLKS
-    v_task_count = b * h_kv * K_NUM_BLKS
-
-    grid = (q_task_count + k_task_count + v_task_count,)
-
-    # call sage_quant_kernel
-    sage_quant_kernel[grid](
-        q,
-        q_int8,
-        q_scale,
-        k,
-        k_int8,
-        k_scale,
-        v,
-        v_fp8,
-        v_scale,
-        stride_bz_q,
-        stride_h_q,
-        stride_seq_q,
-        stride_bz_k,
-        stride_h_k,
-        stride_seq_k,
-        q_scale.stride(0),
-        q_scale.stride(1),
-        k_scale.stride(0),
-        k_scale.stride(1),
-        v_scale.stride(0),
-        v_scale.stride(1),
-        (sm_scale * 1.4426950408889634),
-        q_task_count,
-        k_task_count,
-        b,
-        h_qo,
-        h_kv,
-        Q_NUM_BLKS,
-        K_NUM_BLKS,
-        qo_len,
-        kv_len,
-        triton.next_power_of_2(kv_len),
-        FP8_MAX=FP8_MAX,
-        INT8_MAX=torch.iinfo(q_int8.dtype).max,
-        D=head_dim,
-        BLK_Q=BLKQ,
-        BLK_K=BLKK,
-        num_stages=3,
-        num_warps=8,
-    )
-
-    return q_int8, q_scale, k_int8, k_scale, v_fp8, v_scale
-
-
-@triton.jit
-def sage_quant_kernel(
-    Q_Input,
-    Q_Output,
-    Q_Scale,
-    K_Input,
-    K_Output,
-    K_Scale,
-    V_Input,
-    V_Output,
-    V_Scale,
-    stride_qz,
-    stride_qh,
-    stride_qn,
-    stride_kz,
-    stride_kh,
-    stride_kn,
-    stride_qsz,
-    stride_qsh,
-    stride_ksz,
-    stride_ksh,
-    stride_vsz,
-    stride_vsh,
-    sm_scale,
-    q_task_count,
-    k_task_count,
-    BATCH,
-    Q_HEAD,
-    K_HEAD,
-    Q_NUM_BLKS,
-    K_NUM_BLKS,
-    SEQLEN_Q,
-    SEQLEN_K,
-    SEQLEN_K_PADDED: tl.constexpr,
-    FP8_MAX: tl.constexpr,
-    INT8_MAX: tl.constexpr,
-    D: tl.constexpr,
-    BLK_Q: tl.constexpr,
-    BLK_K: tl.constexpr,
-):
-    pid = tl.program_id(0).to(tl.int64)
-
-    offs_blk_q = tl.arange(0, BLK_Q)
-    offs_blk_k = tl.arange(0, BLK_K)
-    offs_d = tl.arange(0, D)
-
-    if pid < q_task_count:
-        # here we do Q
-        off_blk, off_h, off_b = pid_grid_3d(pid, Q_NUM_BLKS, Q_HEAD, BATCH)
-        offs_qn = off_blk * BLK_Q + offs_blk_q
-
-        q_offs = (
-            off_b * stride_qz
-            + off_h * stride_qh
-            + offs_qn[:, None] * stride_qn
-            + offs_d[None, :]
-        )
-
-        q_input_ptrs = Q_Input + q_offs
-        q_output_ptrs = Q_Output + q_offs
-        q_scale_ptrs = Q_Scale + off_b * stride_qsz + off_h * stride_qsh + off_blk
-
-        _general_quant_kernel(
-            q_input_ptrs,
-            q_output_ptrs,
-            q_scale_ptrs,
-            INT8_MAX,
-            offs_qn[:, None] < SEQLEN_Q,
-            sm_scale=sm_scale,
-        )
-    elif pid >= q_task_count and pid < q_task_count + k_task_count:
-        # here we do K
-        _pid = pid - q_task_count
-        off_blk, off_h, off_b = pid_grid_3d(_pid, K_NUM_BLKS, K_HEAD, BATCH)
-
-        offs_kn = off_blk * BLK_K + offs_blk_k
-
-        k_offs = (
-            off_b * stride_kz
-            + off_h * stride_kh
-            + offs_kn[:, None] * stride_kn
-            + offs_d[None, :]
-        )
-
-        k_input_ptrs = K_Input + k_offs
-        k_output_ptrs = K_Output + k_offs
-        k_scale_ptrs = K_Scale + off_b * stride_ksz + off_h * stride_ksh + off_blk
-
-        _general_quant_kernel(
-            k_input_ptrs,
-            k_output_ptrs,
-            k_scale_ptrs,
-            INT8_MAX,
-            offs_kn[:, None] < SEQLEN_K,
-        )
-    else:
-        # V
-        _pid = pid - (q_task_count + k_task_count)
-        off_blk, off_h, off_b = pid_grid_3d(_pid, K_NUM_BLKS, K_HEAD, BATCH)
-        offs_kn = off_blk * BLK_K + offs_blk_k
-
-        v_offs = (
-            off_b * stride_kz
-            + off_h * stride_kh
-            + offs_kn[:, None] * stride_kn
-            + offs_d[None, :]
-        )
-
-        v_input_ptrs = V_Input + v_offs
-        v_output_ptrs = V_Output + v_offs
-
-        # just apply the per channel v_scales that have been computed outside
-        v_scale_ptrs = (
-            V_Scale + off_b * stride_vsz + off_h * stride_vsh + offs_d[None, :]
-        )
-        v = tl.load(v_input_ptrs, mask=offs_kn[:, None] < SEQLEN_K, other=0.0)
-        v = v.to(tl.float32)
-        v_scales = tl.load(v_scale_ptrs)
-        v_quant = v / v_scales
-        v_quant = v_quant.to(v_output_ptrs.dtype.element_ty)
-        tl.store(v_output_ptrs, v_quant, mask=offs_kn[:, None] < SEQLEN_K)
-
-
-@triton.jit
-def _general_quant_kernel(
-    input_ptrs, output_ptrs, scale_ptrs, DTYPE_MAX, mask, sm_scale=None
-):
-    if mask is not None:
-        x = tl.load(input_ptrs, mask=mask, other=0.0)
-    else:
-        x = tl.load(input_ptrs)
-    x = x.to(tl.float32)
-    if sm_scale is not None:
-        x *= sm_scale
-    scale = tl.max(tl.abs(x)) / DTYPE_MAX
-    x_quant = x / scale
-    if output_ptrs.dtype.element_ty == tl.int8:
-        x_quant += 0.5 * tl.where(x_quant >= 0, 1, -1)
-    x_quant = x_quant.to(output_ptrs.dtype.element_ty)
-    tl.store(output_ptrs, x_quant, mask=mask)
-    tl.store(scale_ptrs, scale)
