+++ aiter/ops/triton/_triton_kernels/attention/fav3_sage_attention_mxfp4.py
@@ -11,8 +11,6 @@ from aiter.ops.triton._triton_kernels.attention.fav3_sage_attention import (
     map_dims,
 )
 
-from aiter.ops.triton.moe.quant_moe import downcast_to_mxfp
-
 
 @triton.jit
 def compute_padding_info(seqlen_k, BLOCK_N: tl.constexpr):
@@ -612,1110 +610,3 @@ def sage_fwd_mxfp4(
     if PADDED_HEAD_V:
         o_mask &= offs_d_v[None, :] < ACTUAL_BLOCK_DMODEL_V
     tl.store(o_ptr, acc.to(Out.dtype.element_ty), mask=o_mask)
-
-
-@triton.jit
-def _compute_mx_quant_and_scale(
-    src_tensor,
-    valid_src_mask,
-    mx_tensor_dtype: tl.constexpr,
-):
-    """
-    Compute MX quantization with RNE (Round to Nearest Even) rounding for the scale.
-
-    RNE is applied when converting max_abs to E8M0 format (nearest power of 2).
-    This is equivalent to computing: scale = 2^(clip(floor(log2(RNE(max_abs(x)))), -127, 127) - 2)
-    where RNE rounds to the nearest power of 2, with ties going to even exponent.
-    """
-    is_fp8: tl.constexpr = (
-        mx_tensor_dtype == tl.float8e4nv or mx_tensor_dtype == tl.float8e5
-    )
-    BLOCK_SIZE_OUT_DIM: tl.constexpr = src_tensor.shape[0]
-    BLOCK_SIZE_QUANT_DIM: tl.constexpr = src_tensor.shape[1]
-    BLOCK_SIZE_QUANT_MX_SCALE: tl.constexpr = src_tensor.shape[1] // 32
-
-    # Explicit cast to fp32 since most ops are not supported on bfloat16
-    f32_tensor = src_tensor.to(tl.float32)
-    abs_tensor = tl.abs(f32_tensor)
-    abs_tensor = tl.where(
-        valid_src_mask, abs_tensor, -1.0
-    )  # Don't consider padding tensors in scale computation
-    abs_tensor = tl.reshape(
-        abs_tensor, [BLOCK_SIZE_OUT_DIM, BLOCK_SIZE_QUANT_MX_SCALE, 32]
-    )
-    max_val = tl.max(abs_tensor, axis=2, keep_dims=True)
-
-    # RNE (Round to Nearest Even) rounding when converting max_abs to E8M0 format
-    # E8M0 stores only exponent (no mantissa), so we round to nearest power of 2
-    # Extract exponent and mantissa from float32
-    max_val_bits = max_val.to(tl.uint32, bitcast=True)
-    exponent = (max_val_bits >> 23) & 0xFF
-    mantissa = max_val_bits & 0x7FFFFF
-
-    # RNE to nearest power of 2:
-    # For value 2^n * (1 + m/2^23), the threshold is at m = 0.5 * 2^23 = 0x400000
-    # - If mantissa < 0x400000: round to 2^n (keep exponent)
-    # - If mantissa > 0x400000: round to 2^(n+1) (increment exponent)
-    # - If mantissa == 0x400000: tie case, round to even exponent (RNE)
-
-    # Determine if we should round up
-    should_round_up = (mantissa > 0x400000) | (
-        (mantissa == 0x400000) & ((exponent & 1) == 1)
-    )
-
-    rounded_exponent = tl.where(should_round_up, exponent + 1, exponent)
-
-    # Subtract 2 from exponent (divide by 4) to get final scale exponent
-    # Clamp to valid E8M0 range [-127, 127] (exponent 0-254 in biased representation)
-    scale_exponent = rounded_exponent - 2
-    scale_exponent = tl.maximum(scale_exponent, 0)
-    scale_exponent = tl.minimum(scale_exponent, 254)
-
-    # Construct the scale as a power of 2
-    dequant_scale_exponent = (scale_exponent << 23) & 0x7F800000
-    dequant_scale = dequant_scale_exponent.to(tl.float32, bitcast=True)
-    quant_scale = tl.where(dequant_scale == 0, 0, 1.0 / dequant_scale)
-
-    f32_tensor = tl.reshape(
-        f32_tensor, [BLOCK_SIZE_OUT_DIM, BLOCK_SIZE_QUANT_MX_SCALE, 32]
-    )
-    quant_tensor = f32_tensor * quant_scale
-
-    # Reshape the tensors after scaling
-    quant_tensor = quant_tensor.reshape([BLOCK_SIZE_OUT_DIM, BLOCK_SIZE_QUANT_DIM])
-    # Set the invalid portions of the tensor to 0
-    quant_tensor = tl.where(valid_src_mask, quant_tensor, 0)
-    dequant_scale_exponent = dequant_scale_exponent.reshape(
-        [BLOCK_SIZE_OUT_DIM, BLOCK_SIZE_QUANT_MX_SCALE]
-    )
-
-    # Extract the exponent part of the scales and store the result
-    dequant_scale_exponent = (dequant_scale_exponent >> 23).to(tl.uint8)
-
-    # Convert the tensors to the mx format
-    if is_fp8:
-        out_tensor = quant_tensor.to(mx_tensor_dtype)
-    else:
-        quant_tensor = quant_tensor.to(tl.uint32, bitcast=True)
-        signs = quant_tensor & 0x80000000
-        exponents = (quant_tensor >> 23) & 0xFF
-        mantissas = quant_tensor & 0x7FFFFF
-
-        # 0.25 <= x < 0.75 maps to 0.5, a denormal number
-        E8_BIAS = 127
-        E2_BIAS = 1
-        # Move implicit bit 1 at the beginning to mantissa for denormals
-        adjusted_exponents = tl.core.sub(
-            E8_BIAS, exponents + 1, sanitize_overflow=False
-        )
-        mantissas = tl.where(
-            exponents < E8_BIAS,
-            (0x400000 | (mantissas >> 1)) >> adjusted_exponents,
-            mantissas,
-        )
-
-        # For normal numbers, we change the bias from 127 to 1, and for subnormals, we keep exponent as 0.
-        exponents = tl.maximum(exponents, E8_BIAS - E2_BIAS) - (E8_BIAS - E2_BIAS)
-
-        # Combine sign, exponent, and mantissa, while saturating
-        # rounding nearest with tie breaking up by adding +1 to one bit right of the LSB, then shift right
-        e2m1_tmp = tl.minimum((((exponents << 2) | (mantissas >> 21)) + 1) >> 1, 0x7)
-        e2m1_value = ((signs >> 28) | e2m1_tmp).to(tl.uint8)
-
-        e2m1_value = tl.reshape(
-            e2m1_value, [BLOCK_SIZE_OUT_DIM, BLOCK_SIZE_QUANT_DIM // 2, 2]
-        )
-        evens, odds = tl.split(e2m1_value)
-        out_tensor = evens | (odds << 4)
-
-    return out_tensor, dequant_scale_exponent
-
-
-def fused_sage_quant_mxfp4(
-    q,
-    k,
-    v,
-    BLOCK_M,
-    hadamard_rotation=False,
-    R=None,
-    BLOCK_R=None,
-    q_smoothing=False,
-    layout="bshd",
-):
-
-    if layout == "bhsd":
-        b, h_qo, qo_len, head_dim = q.shape
-        _, h_kv, kv_len, _ = v.shape
-
-        stride_bz_v, stride_h_v, stride_seq_v, stride_d_v = v.stride(0), v.stride(1), v.stride(2), v.stride(3)
-
-    elif layout == "bshd":
-        b, qo_len, h_qo, head_dim = q.shape
-        _, kv_len, h_kv, _ = v.shape
-
-        stride_bz_v, stride_h_v, stride_seq_v, stride_d_v = v.stride(0), v.stride(2), v.stride(1), v.stride(3)
-    else:
-        raise ValueError(f"Unknown tensor layout: {layout}")
-
-    # padded_head_dim = max(16, 1 << (head_dim - 1).bit_length())
-    sm_scale = head_dim**-0.5
-    
-    q_fp4, q_scale, k_fp4, k_scale, delta_s = smooth_rotate_downcast_qk(
-        q,
-        k,
-        BLOCK_SIZE_M=BLOCK_M,
-        hadamard_rotation=hadamard_rotation,
-        R=R,
-        BLOCK_R=BLOCK_R,
-        q_smoothing=q_smoothing,
-        layout=layout,
-        sm_scale=(sm_scale * 1.4426950408889634),
-    )
-
-    FP8_TYPE = aiter.dtypes.fp8
-    FP8_MAX = torch.finfo(FP8_TYPE).max
-    v_fp8 = torch.empty_like(v, dtype=FP8_TYPE, device=v.device)
-
-    BLOCK_K = 1024
-    K_NUM_BLKS = (kv_len + BLOCK_K - 1) // BLOCK_K
-
-    # Apply K tensor smoothing following SageAttention approach
-    v_scale = v.abs().amax(dim=1 if layout == "bshd" else 2).to(torch.float32) / FP8_MAX
-
-    v_task_count = b * h_kv * K_NUM_BLKS
-    grid = (v_task_count,)
-    sage_quant_v_kernel[grid](
-        v,
-        v_fp8,
-        v_scale,
-        stride_bz_v,
-        stride_h_v,
-        stride_seq_v,
-        stride_d_v,
-        v_scale.stride(0),
-        v_scale.stride(1),
-        b,
-        h_kv,
-        K_NUM_BLKS,
-        kv_len,
-        D=head_dim,
-        BLK_K=BLOCK_K,
-        num_stages=5,
-        num_warps=8,
-    )
-
-    return q_fp4, q_scale, k_fp4, k_scale, v_fp8, v_scale, delta_s
-
-
-@triton.jit
-def sage_quant_v_kernel(
-    V_Input,
-    V_Output,
-    V_Scale,
-    stride_kz,
-    stride_kh,
-    stride_kn,
-    stride_kd,
-    stride_vsz,
-    stride_vsh,
-    BATCH,
-    K_HEAD,
-    K_NUM_BLKS,
-    SEQLEN_K,
-    D: tl.constexpr,
-    BLK_K: tl.constexpr,
-):
-    pid = tl.program_id(0).to(tl.int64)
-
-    offs_blk_k = tl.arange(0, BLK_K)
-    offs_d = tl.arange(0, D)
-
-    # V
-    off_blk, off_h, off_b = pid_grid_3d(pid, K_NUM_BLKS, K_HEAD, BATCH)
-    offs_kn = off_blk * BLK_K + offs_blk_k
-
-    v_offs = (
-        off_b * stride_kz
-        + off_h * stride_kh
-        + offs_kn[:, None] * stride_kn
-        + offs_d[None, :] * stride_kd
-    )
-
-    v_input_ptrs = V_Input + v_offs
-    v_output_ptrs = V_Output + v_offs
-
-    # just apply the per channel v_scales that have been computed outside
-    v_scale_ptrs = V_Scale + off_b * stride_vsz + off_h * stride_vsh + offs_d[None, :]
-    v = tl.load(v_input_ptrs, mask=offs_kn[:, None] < SEQLEN_K, other=0.0)
-    v = v.to(tl.float32)
-    v_scales = tl.load(v_scale_ptrs)
-    v_quant = v / v_scales
-    v_quant = v_quant.to(v_output_ptrs.dtype.element_ty)
-    tl.store(v_output_ptrs, v_quant, mask=offs_kn[:, None] < SEQLEN_K)
-
-
-@triton.jit
-def _rotate_quantize_qk_kernel(
-    Q,
-    Q_q,
-    Q_descale,
-    Q_mean,
-    K,
-    K_q,
-    K_descale,
-    R,  # Hadamard matrix
-    sm_scale: tl.constexpr,
-    stride_qb,
-    stride_qh,
-    stride_qm,
-    stride_qd,
-    stride_qqb,
-    stride_qqm,
-    stride_qqh,
-    stride_qqd,
-    stride_qsb,
-    stride_qsm,
-    stride_qsh,
-    stride_qsd,
-    stride_mb,
-    stride_mh,
-    stride_mm,
-    stride_md,
-    stride_kb,
-    stride_kh,
-    stride_km,
-    stride_kd,
-    stride_kqb,
-    stride_kqn,
-    stride_kqh,
-    stride_kqd,
-    stride_ksb,
-    stride_ksn,
-    stride_ksh,
-    stride_ksd,
-    batch,
-    heads_q,
-    heads_k,
-    seqlen_q,
-    seqlen_k,
-    d_model,
-    q_smoothing: tl.constexpr,
-    hadamard_rotation: tl.constexpr,
-    BLOCK_M: tl.constexpr,
-    BLOCK_R: tl.constexpr,  # rotation block size
-    D: tl.constexpr,  # D is 128
-):
-    SCALE_GROUP_SIZE: tl.constexpr = 32
-
-    q_pids = batch * heads_q * tl.cdiv(seqlen_q, BLOCK_M)
-    pid = tl.program_id(0).to(tl.int64)
-    is_q_pid = pid < q_pids
-
-    if is_q_pid:
-        pid_b = pid % batch
-        pid_h = pid // batch % heads_q
-        pid_m = pid // (batch * heads_q)
-    else:  # is k pid
-        pid -= q_pids
-        pid_b = pid % batch
-        pid_h = pid // batch % heads_k
-        pid_m = pid // (batch * heads_k)
-
-    # Offsets
-    offs_m = pid_m * BLOCK_M + tl.arange(0, BLOCK_M)
-    offs_d = tl.arange(0, D)
-
-    offs_dq = tl.arange(0, D // 2)
-    offs_ds = tl.arange(0, D // SCALE_GROUP_SIZE)
-
-    # set pointers to either Q or K tensor, descale, quantized output
-    # Q block shape: [BLOCK_M, D]
-    if is_q_pid:
-        tensor_offset = Q + (
-            pid_b * stride_qb
-            + pid_h * stride_qh
-            + offs_m[:, None] * stride_qm
-            + offs_d[None, :] * stride_qd
-        )
-        descale_offset = Q_descale + (
-            pid_b * stride_qsb
-            + pid_h * stride_qsh
-            + offs_m[:, None] * stride_qsm
-            + offs_ds[None, :] * stride_qsd
-        )  # we group 32 values together for quantization
-        # Store rotated and quantized Q
-        quant_tensor_offset = Q_q + (
-            pid_b * stride_qqb
-            + pid_h * stride_qqh
-            + offs_m[:, None] * stride_qqm
-            + offs_dq[None, :] * stride_qqd
-        )
-        seqlen = seqlen_q
-    else:
-        tensor_offset = K + (
-            pid_b * stride_kb
-            + pid_h * stride_kh
-            + offs_m[:, None] * stride_km
-            + offs_d[None, :] * stride_kd
-        )
-        descale_offset = K_descale + (
-            pid_b * stride_ksb
-            + pid_h * stride_ksh
-            + offs_m[:, None] * stride_ksn
-            + offs_ds[None, :] * stride_ksd
-        )  # we group 32 values together for quantization
-
-        quant_tensor_offset = K_q + (
-            pid_b * stride_kqb
-            + pid_h * stride_kqh
-            + offs_m[:, None] * stride_kqn
-            + offs_dq[None, :] * stride_kqd
-        )
-        seqlen = seqlen_k
-
-    qk_ptr = tensor_offset
-    qk_descale_ptr = descale_offset
-    qk_quant_ptr = quant_tensor_offset
-
-    qk_tile = tl.load(
-        qk_ptr, mask=(offs_m[:, None] < seqlen) & (offs_d[None, :] < d_model), other=0.0
-    )  # (BLOCK_M, D)
-    original_dtype = qk_tile.dtype
-
-    if is_q_pid:
-        if q_smoothing:
-            ACTUAL_BLOCK_M = tl.minimum(BLOCK_M, seqlen - pid_m * BLOCK_M)
-            m_row_mean = (
-                tl.sum(qk_tile, axis=0) / ACTUAL_BLOCK_M
-            )  # Sum over BLOCK_M -> shape [D]
-            qk_tile -= m_row_mean[None, :]
-            qk_tile = qk_tile.to(original_dtype)
-            mean_ptr = (
-                Q_mean
-                + pid_b * stride_mb
-                + pid_h * stride_mh
-                + pid_m * stride_mm
-                + offs_d * stride_md
-            )
-            tl.store(mean_ptr, m_row_mean * sm_scale)
-
-    if hadamard_rotation:
-        r_ptr = (
-            R
-            + tl.arange(0, BLOCK_R)[:, None] * BLOCK_R
-            + tl.arange(0, BLOCK_R)[None, :]
-        )
-        r_mat = tl.load(r_ptr)  # BLOCK_R x BLOCK_R
-
-        shape0: tl.constexpr = BLOCK_M * D // BLOCK_R
-
-        # Rotate: Q_rot = Q @ R
-        qk_rot_tile = tl.dot(qk_tile.reshape((shape0, BLOCK_R)).to(r_mat.dtype), r_mat)
-        qk_rot_tile = qk_rot_tile.reshape((BLOCK_M, D))
-    else:
-        qk_rot_tile = qk_tile.to(tl.float32)
-
-    if is_q_pid:
-        qk_rot_tile *= sm_scale
-
-    qk_quant_tile, qk_descale = _compute_mx_quant_and_scale(
-        qk_rot_tile, offs_m[:, None] < seqlen, tl.uint8
-    )
-
-    tl.store(qk_descale_ptr, qk_descale, mask=(offs_m[:, None] < seqlen))
-
-    tl.store(
-        qk_quant_ptr,
-        qk_quant_tile,
-        mask=(offs_m[:, None] < seqlen),
-    )
-
-
-@triton.jit
-def _compute_delta_s_kernel(
-    Q_mean,
-    K_rot,
-    Delta_S,
-    stride_mb,
-    stride_mh,
-    stride_mm,
-    stride_md,
-    stride_kb,
-    stride_kh,
-    stride_kn,
-    stride_kd,
-    stride_sb,
-    stride_sh,
-    stride_sm,
-    stride_sn,
-    n_heads_k,
-    n_heads_q,
-    seq_k,
-    d_model,
-    BLOCK_N: tl.constexpr,  # Number of K-tokens to process
-):
-    pid_bh = tl.program_id(0)
-    pid_m_q = tl.program_id(1)  # The Q-block index
-    pid_n_k = tl.program_id(2)  # The K-block index
-
-    pid_hq = pid_bh % n_heads_q
-    pid_h = pid_hq // (n_heads_q // n_heads_k)
-    pid_b = pid_bh // n_heads_q
-
-    offs_n = pid_n_k * BLOCK_N + tl.arange(0, BLOCK_N)
-
-    # Accumulate dot product across the whole d_model
-    acc = tl.zeros([BLOCK_N], dtype=tl.float32)
-
-    # Loop over d_model in steps of 32 (our block_size)
-    for d_offset in range(0, d_model, 32):
-        offs_d = d_offset + tl.arange(0, 32)
-
-        # Load Q_mean segment: [32]
-        qm_ptr = (
-            Q_mean
-            + pid_b * stride_mb
-            + pid_hq * stride_mh
-            + pid_m_q * stride_mm
-            + offs_d * stride_md
-        )
-        qm_val = tl.load(qm_ptr)
-
-        # Load K_rot segment: [BLOCK_N, 32]
-        kn_ptr = (
-            K_rot
-            + pid_b * stride_kb
-            + pid_h * stride_kh
-            + offs_n[:, None] * stride_kn
-            + offs_d[None, :] * stride_kd
-        )
-        kn_val = tl.load(kn_ptr, mask=offs_n[:, None] < seq_k, other=0.0)
-
-        # Compute dot product for this d-segment
-        acc += tl.sum(qm_val[None, :] * kn_val, axis=1)
-
-    # Store to Delta_S [B, H, Q_BLKS, seq_k]
-    s_ptr = (
-        Delta_S
-        + pid_b * stride_sb
-        + pid_hq * stride_sh
-        + pid_m_q * stride_sm
-        + offs_n * stride_sn
-    )
-    tl.store(s_ptr, acc, mask=offs_n < seq_k)
-
-
-# @functools.lru_cache(maxsize=16)
-def create_hadamard_matrix(block_size, device="cuda", dtype=torch.float32):
-    """
-    Returns a Hadamard matrix of size block_size x block_size. Remember to normalize with sqrt(block_size) for it to be orthogonal.
-    """
-    assert (block_size & (block_size - 1)) == 0, "block_size must be power of 2"
-    assert block_size > 0, "block_size must be positive"
-
-    # Base case: H_1 = [1]
-    if block_size == 1:
-        return torch.ones(1, 1, device=device, dtype=dtype)
-
-    # Recursive construction: H_{2n} = [H_n   H_n  ]
-    #                                   [H_n  -H_n ]
-    H_half = create_hadamard_matrix(block_size // 2, device=device, dtype=dtype)
-
-    # Build the full matrix (unnormalized)
-    H = torch.zeros(block_size, block_size, device=device, dtype=dtype)
-    half = block_size // 2
-    H[:half, :half] = H_half
-    H[:half, half:] = H_half
-    H[half:, :half] = H_half
-    H[half:, half:] = -H_half
-
-    # The unnormalized matrix satisfies H_unnorm @ H_unnorm.T = block_size * I
-    # remember to divide by sqrt(block_size) to get orthogonal matrix
-    return H
-
-
-def smooth_rotate_downcast_qk(
-    q,
-    k,
-    BLOCK_SIZE_M,
-    hadamard_rotation=False,
-    R=None,
-    BLOCK_R=None,
-    q_smoothing=False,
-    sm_scale=None,
-    layout="bhsd",
-):
-    if hadamard_rotation:
-        if R is None:
-            assert (
-                BLOCK_R is not None
-            ), "if using hadamard rotation, BLOCK_R (size of the hadamard matrix) must be provided."
-            R = create_hadamard_matrix(BLOCK_R, q.device) / (BLOCK_R**0.5)
-        else:
-            BLOCK_R = R.shape[-1]
-
-    bshd = [0, 1, 2, 3] if layout == "bshd" else [0, 2, 1, 3]
-
-    # shapes
-    b, s_q, h_q, d = map_dims(q.shape, bshd)
-    _, s_k, h_k, _ = map_dims(k.shape, bshd)
-
-    Q_NUM_BLKS = (s_q + BLOCK_SIZE_M - 1) // BLOCK_SIZE_M
-    K_NUM_BLKS = (s_k + BLOCK_SIZE_M - 1) // BLOCK_SIZE_M
-
-    if q_smoothing:
-        q_mean = torch.empty(
-            (b, h_q, Q_NUM_BLKS, d), dtype=torch.float32, device=q.device
-        )
-        delta_s = torch.empty(
-            (b, h_q, Q_NUM_BLKS, s_k), dtype=torch.float32, device=q.device
-        )
-    else:
-        q_mean = None
-        delta_s = None
-
-    stride_qb, stride_qm, stride_qh, stride_qd = map_dims(q.stride(), bshd)
-    stride_kb, stride_kn, stride_kh, stride_kd = map_dims(k.stride(), bshd)
-
-    Q_q = torch.empty((*q.shape[:-1], d // 2), dtype=torch.uint8, device=q.device)
-    Q_descale = torch.empty(
-        (*q.shape[:-1], d // 32), dtype=torch.uint8, device=q.device
-    )
-    K_q = torch.empty((*k.shape[:-1], d // 2), dtype=torch.uint8, device=k.device)
-    K_descale = torch.empty(
-        (*k.shape[:-1], d // 32), dtype=torch.uint8, device=k.device
-    )
-
-    stride_qqb, stride_qqm, stride_qqh, stride_qqd = map_dims(Q_q.stride(), bshd)
-    stride_kqb, stride_kqn, stride_kqh, stride_kqd = map_dims(K_q.stride(), bshd)
-
-    stride_qsb, stride_qsm, stride_qsh, stride_qsd = map_dims(Q_descale.stride(), bshd)
-    stride_ksb, stride_ksn, stride_ksh, stride_ksd = map_dims(K_descale.stride(), bshd)
-
-    grid = (b * (h_q * Q_NUM_BLKS + h_k * K_NUM_BLKS),)
-    _rotate_quantize_qk_kernel[grid](
-        q,
-        Q_q,
-        Q_descale,
-        q_mean,
-        k,
-        K_q,
-        K_descale,
-        R,
-        sm_scale,
-        stride_qb,
-        stride_qh,
-        stride_qm,
-        stride_qd,
-        stride_qqb,
-        stride_qqm,
-        stride_qqh,
-        stride_qqd,
-        stride_qsb,
-        stride_qsm,
-        stride_qsh,
-        stride_qsd,
-        q_mean.stride(0) if q_smoothing else None,
-        q_mean.stride(1) if q_smoothing else None,
-        q_mean.stride(2) if q_smoothing else None,
-        q_mean.stride(3) if q_smoothing else None,
-        stride_kb,
-        stride_kh,
-        stride_kn,
-        stride_kd,
-        stride_kqb,
-        stride_kqn,
-        stride_kqh,
-        stride_kqd,
-        stride_ksb,
-        stride_ksn,
-        stride_ksh,
-        stride_ksd,
-        b,
-        h_q,
-        h_k,
-        s_q,
-        s_k,
-        d,
-        q_smoothing=q_smoothing,
-        hadamard_rotation=hadamard_rotation,
-        BLOCK_M=BLOCK_SIZE_M,
-        BLOCK_R=BLOCK_R,
-        D=d,
-        num_warps=4,
-        num_stages=5
-    )
-
-    if q_smoothing:
-        # 3. Compute Smoothing Delta S
-        # Grid: Each Q-block x Each K-block
-        grid_delta = (b * h_q, Q_NUM_BLKS, K_NUM_BLKS)
-        _compute_delta_s_kernel[grid_delta](
-            q_mean,
-            k,
-            delta_s,
-            q_mean.stride(0),
-            q_mean.stride(1),
-            q_mean.stride(2),
-            q_mean.stride(3),
-            stride_kb,
-            stride_kh,
-            stride_kn,
-            stride_kd,
-            delta_s.stride(0),
-            delta_s.stride(1),
-            delta_s.stride(2),
-            delta_s.stride(3),
-            h_k,
-            h_q,
-            s_k,
-            d,
-            BLOCK_N=BLOCK_SIZE_M,
-        )
-
-    return Q_q, Q_descale, K_q, K_descale, delta_s
-
-
-
-def sage_quant_mxfp4(
-    q,
-    k,
-    v,
-    FP8_TYPE,
-    FP8_MAX,
-    BLKQ=128,
-    BLKK=64,
-    sm_scale=None,
-    q_smoothing=False,
-    layout="bshd",
-    USE_RNE=False,
-):
-    v_fp8 = torch.empty_like(v, dtype=FP8_TYPE, device=v.device)
-
-    if layout == "bhsd":
-        b, h_qo, qo_len, head_dim = q.shape
-        _, h_kv, kv_len, _ = v.shape
-
-        stride_bz_v, stride_h_v, stride_seq_v, stride_d_v= v.stride(0), v.stride(1), v.stride(2), v.stride(3)
-
-    elif layout == "bshd":
-        b, qo_len, h_qo, head_dim = q.shape
-        _, kv_len, h_kv, _ = v.shape
-
-        stride_bz_v, stride_h_v, stride_seq_v, stride_d_v = v.stride(0), v.stride(2), v.stride(1), v.stride(3)
-    else:
-        raise ValueError(f"Unknown tensor layout: {layout}")
-    K_NUM_BLKS = (kv_len + BLKK - 1) // BLKK
-
-    # Apply K tensor smoothing following SageAttention approach
-    v_scale = v.abs().amax(dim=1 if layout == "bshd" else 2).to(torch.float32) / FP8_MAX
-
-    v_task_count = b * h_kv * K_NUM_BLKS
-    grid = (v_task_count,)
-
-    padded_head_dim = max(16, 1 << (head_dim - 1).bit_length())
-
-    if sm_scale is None:
-        sm_scale = head_dim**-0.5
-    
-    q, k, delta_s = rotation_smooth_qk(q, k, BLKQ, block_size=padded_head_dim, q_smoothing=q_smoothing, layout=layout, sm_scale=(sm_scale * 1.4426950408889634))
-
-
-    sage_quant_v_kernel[grid](
-        v,
-        v_fp8,
-        v_scale,
-        stride_bz_v,
-        stride_h_v,
-        stride_seq_v,
-        stride_d_v,
-        v_scale.stride(0),
-        v_scale.stride(1),
-        b,
-        h_kv,
-        K_NUM_BLKS,
-        kv_len,
-        D=head_dim,
-        BLK_K=BLKK,
-        num_stages=3,
-        num_warps=8,
-    )
-
-    downcast_func = downcast_to_mxfp
-
-    q_fp4, q_scale = downcast_func(q, torch.uint8, axis=-1)
-    k_fp4, k_scale = downcast_func(k, torch.uint8, axis=-1)
-
-    return q_fp4, q_scale, k_fp4, k_scale, v_fp8, v_scale, delta_s
-
-
-@triton.jit
-def _rot_q_kernel(
-    Q,
-    Q_rot,
-    Q_mean,
-    R,  # Hadamard matrix
-    sm_scale: tl.constexpr,
-    stride_qb,
-    stride_qh,
-    stride_qm,
-    stride_qd,
-    stride_qob,
-    stride_qoh,
-    stride_qom,
-    stride_qod,
-    stride_mb,
-    stride_mh,
-    stride_mm,
-    stride_md,
-    stride_rm,
-    stride_rd,
-    n_heads,
-    seq_len,
-    d_model,
-    q_smoothing: tl.constexpr,
-    BLOCK_M: tl.constexpr,
-    BLOCK_D: tl.constexpr,  # BLOCK_D is 32
-):
-    # Grid: (batch * n_heads, seq_len // BLOCK_M, d_model // BLOCK_D)
-    pid_bh = tl.program_id(0).to(tl.int64)
-    pid_m = tl.program_id(1).to(tl.int64)
-    pid_d = tl.program_id(2).to(tl.int64)
-
-    pid_h = pid_bh % n_heads
-    pid_b = pid_bh // n_heads
-
-    # Offsets
-    offs_m = pid_m * BLOCK_M + tl.arange(0, BLOCK_M)
-    offs_d = pid_d * BLOCK_D + tl.arange(0, BLOCK_D)
-
-    # Load Q block and R (Hadamard)
-    # Q block shape: [BLOCK_M, BLOCK_D]
-    q_ptr = (
-        Q
-        + (pid_b * stride_qb)
-        + pid_h * stride_qh
-        + offs_m[:, None] * stride_qm
-        + offs_d[None, :] * stride_qd
-    )
-    r_ptr = (
-        R + tl.arange(0, BLOCK_D)[:, None] * stride_rm + tl.arange(0, BLOCK_D)[None, :] * stride_rd
-    )
-    q_tile = tl.load(
-        q_ptr, mask=(offs_m[:, None] < seq_len) & (offs_d[None, :] < d_model), other=0.0
-    )
-    r_mat = tl.load(r_ptr)  # 32x32
-
-    # Rotate: Q_rot = Q @ R
-    q_rot_tile = tl.dot(q_tile, r_mat)
-    if sm_scale is not None:
-        q_rot_tile *= sm_scale
-
-    # Store rotated Q
-    rot_ptr = (
-        Q_rot
-        + (pid_b * stride_qob)
-        + pid_h * stride_qoh
-        + offs_m[:, None] * stride_qom
-        + offs_d[None, :] * stride_qod
-    )
-
-    # Calculate mean for the block (reduction over d within the BLOCK_M)
-    # q_mean shape: [B, H, Q_NUM_BLKS, D]
-    if q_smoothing:
-        m_row_mean = (
-            tl.sum(q_rot_tile, axis=0) / BLOCK_M
-        )  # Sum over BLOCK_M -> shape [BLOCK_D]
-
-        q_rot_tile -= m_row_mean[None, :]
-        # Store mean (Atomic add or structured store)
-        # For simplicity in this layout, we store the block-sum
-        # and divide by BLOCK_M in the host or final step
-        mean_ptr = (
-            Q_mean
-            + (pid_b * stride_mb)
-            + pid_h * stride_mh
-            + pid_m * stride_mm
-            + offs_d * stride_md
-        )
-        tl.store(mean_ptr, m_row_mean)
-
-    tl.store(
-        rot_ptr,
-        q_rot_tile,
-        mask=(offs_m[:, None] < seq_len) & (offs_d[None, :] < d_model),
-    )
-
-
-
-
-@triton.jit
-def _rot_k_only_kernel(
-    K,
-    K_rot,
-    R,
-    stride_kb,
-    stride_kh,
-    stride_kn,
-    stride_kd,
-    stride_kob,
-    stride_koh,
-    stride_kon,
-    stride_kod,
-    stride_rm,
-    stride_rd,
-    n_heads,
-    seq_k,
-    d_model,
-    BLOCK_M: tl.constexpr,
-    BLOCK_D: tl.constexpr,
-):
-    pid_bh = tl.program_id(0).to(tl.int64)
-    pid_n = tl.program_id(1).to(tl.int64)
-    pid_d = tl.program_id(2).to(tl.int64)
-
-    pid_h = pid_bh % n_heads
-    pid_b = pid_bh // n_heads
-
-    offs_n = pid_n * BLOCK_M + tl.arange(0, BLOCK_M)
-    offs_d = pid_d * BLOCK_D + tl.arange(0, BLOCK_D)
-
-    # Load K block and R
-    k_ptr = (
-        K
-        + (pid_b * stride_kb)
-        + (pid_h * stride_kh)
-        + offs_n[:, None] * stride_kn
-        + offs_d[None, :] * stride_kd
-    )
-    r_ptr = (
-        R + tl.arange(0, BLOCK_D)[:, None] * stride_rm + tl.arange(0, BLOCK_D)[None, :] * stride_rd
-    )
-
-    k_tile = tl.load(
-        k_ptr, mask=(offs_n[:, None] < seq_k) & (offs_d[None, :] < d_model), other=0.0
-    )
-    r_mat = tl.load(r_ptr)
-
-    # Rotate K
-    k_rot_tile = tl.dot(k_tile, r_mat)
-
-    # Store
-    rot_ptr = (
-        K_rot
-        + (pid_b * stride_kob)
-        + pid_h * stride_koh
-        + offs_n[:, None] * stride_kon
-        + offs_d[None, :] * stride_kod
-    )
-    tl.store(
-        rot_ptr,
-        k_rot_tile,
-        mask=(offs_n[:, None] < seq_k) & (offs_d[None, :] < d_model),
-    )
-
-
-@triton.jit
-def _compute_delta_s_kernel(
-    Q_mean,
-    K_rot,
-    Delta_S,
-    stride_mb,
-    stride_mh,
-    stride_mm,
-    stride_md,
-    stride_kb,
-    stride_kh,
-    stride_kn,
-    stride_kd,
-    stride_sb,
-    stride_sh,
-    stride_sm,
-    stride_sn,
-    n_heads,
-    seq_k,
-    d_model,
-    BLOCK_N: tl.constexpr,  # Number of K-tokens to process
-):
-    pid_bh = tl.program_id(0).to(tl.int64)
-    pid_m_q = tl.program_id(1).to(tl.int64)  # The Q-block index
-    pid_n_k = tl.program_id(2).to(tl.int64) # The K-block index
-
-    pid_h = pid_bh % n_heads
-    pid_b = pid_bh // n_heads
-
-    offs_n = pid_n_k * BLOCK_N + tl.arange(0, BLOCK_N)
-
-    # Accumulate dot product across the whole d_model
-    acc = tl.zeros([BLOCK_N], dtype=tl.float32)
-
-    # Loop over d_model in steps of 32 (our block_size)
-    for d_offset in range(0, d_model, 32):
-        offs_d = d_offset + tl.arange(0, 32)
-
-        # Load Q_mean segment: [32]
-        qm_ptr = (
-            Q_mean
-            + pid_b * stride_mb
-            + pid_h * stride_mh
-            + pid_m_q * stride_mm
-            + offs_d * stride_md
-        )
-        qm_val = tl.load(qm_ptr)
-
-        # Load K_rot segment: [BLOCK_N, 32]
-        kn_ptr = (
-            K_rot
-            + pid_b * stride_kb
-            + pid_h * stride_kh
-            + offs_n[:, None] * stride_kn
-            + offs_d[None, :] * stride_kd
-        )
-        kn_val = tl.load(kn_ptr, mask=offs_n[:, None] < seq_k, other=0.0)
-
-        # Compute dot product for this d-segment
-        acc += tl.sum(qm_val[None, :] * kn_val, axis=1)
-
-    # Store to Delta_S [B, H, Q_BLKS, seq_k]
-    s_ptr = (
-        Delta_S
-        + pid_b * stride_sb
-        + pid_h * stride_sh
-        + pid_m_q * stride_sm
-        + offs_n * stride_sn
-    )
-    tl.store(s_ptr, acc, mask=offs_n < seq_k)
-
-
-def create_random_hadamard_matrix(block_size, device="cuda", dtype=torch.float32):
-    # 1. Generate the deterministic Hadamard matrix (H)
-    H = create_hadamard_matrix(block_size, dtype=dtype) / (block_size**0.5)
-    # 2. Create the random diagonal matrix D (represented as a vector for efficiency)
-    # This generates random +1 or -1 for each column
-    random_signs = torch.randint(0, 2, (block_size,), device=device, dtype=torch.int) * 2 - 1
-    # 3. Apply the random signs (H @ D)
-    # Multiplying by a diagonal matrix on the right is equivalent to scaling columns
-    H_tilde = H * random_signs
-    return H_tilde
-
-
-def rotation_smooth_qk(q, k, BLOCK_SIZE_M=256, block_size=32, q_smoothing=False, sm_scale=None, layout="bhsd"):
-    # Generate Hadamard Matrix R (Rank 32)
-    # TODO we might want to manually define this matri
-    R = create_hadamard_matrix(block_size, dtype=q.dtype) / (block_size**0.5)
-
-
-    
-    # R = create_random_hadamard_matrix(block_size, dtype=q.dtype)
-    bshd = [0, 1, 2, 3] if layout == "bshd" else [0, 2, 1, 3]
-
-    # shapes
-    b, s_q, h_q, d = map_dims(q.shape, bshd)
-    _, s_k, h_k, _ = map_dims(k.shape, bshd)
-
-    Q_rot = torch.empty_like(q)
-    K_rot = torch.empty_like(k)
-
-    Q_NUM_BLKS = (s_q + BLOCK_SIZE_M - 1) // BLOCK_SIZE_M
-    K_NUM_BLKS = (s_k + BLOCK_SIZE_M - 1) // BLOCK_SIZE_M
-
-    # TODO check the dtypes for scales
-    if q_smoothing:
-        q_mean = torch.empty((b, h_q, Q_NUM_BLKS, d), dtype=torch.float32, device=q.device)
-        delta_s = torch.empty(
-            (b, h_q, Q_NUM_BLKS, s_k), dtype=torch.float32, device=q.device
-        )
-    else:
-        q_mean = None
-        delta_s = None
-
-    stride_qb, stride_qm, stride_qh, stride_qd = map_dims(q.stride(), bshd)
-    stride_qob, stride_qom, stride_qoh, stride_qod = map_dims(Q_rot.stride(), bshd)
-    stride_kb, stride_kn, stride_kh, stride_kd = map_dims(k.stride(), bshd)
-    stride_kob, stride_kon, stride_koh, stride_kod = map_dims(K_rot.stride(), bshd)
-    # Launch Q Kernel
-    grid_q = (b * h_q, Q_NUM_BLKS, d // block_size)
-    _rot_q_kernel[grid_q](
-        q,
-        Q_rot,
-        q_mean,
-        R,
-        sm_scale,
-        stride_qb,
-        stride_qh,
-        stride_qm,
-        stride_qd,
-        stride_qob,
-        stride_qoh,
-        stride_qom,
-        stride_qod,
-        q_mean.stride(0) if q_smoothing else None,
-        q_mean.stride(1) if q_smoothing else None,
-        q_mean.stride(2) if q_smoothing else None,
-        q_mean.stride(3) if q_smoothing else None,
-        R.stride(0),
-        R.stride(1),
-        h_q,
-        s_q,
-        d,
-        q_smoothing=q_smoothing,
-        BLOCK_M=BLOCK_SIZE_M,
-        BLOCK_D=block_size,
-    )
-
-    
-
-    # 2. Rotate K (Only once!)
-    grid_k = (b * h_k, K_NUM_BLKS, d // block_size)
-    _rot_k_only_kernel[grid_k](
-        k,
-        K_rot,
-        R,
-        stride_kb,
-        stride_kh,
-        stride_kn,
-        stride_kd,
-        stride_kob,
-        stride_koh,
-        stride_kon,
-        stride_kod,
-        R.stride(0),
-        R.stride(1),
-        h_k,
-        s_k,
-        d,
-        BLOCK_M=BLOCK_SIZE_M,
-        BLOCK_D=block_size,
-    )
-
-
-    # smooth k after rotation
-    K_rot = K_rot - K_rot.mean(dim=1 if layout == "bshd" else 2, keepdim=True)
-
-    if q_smoothing:
-        # 3. Compute Smoothing Delta S
-        # Grid: Each Q-block x Each K-block
-        grid_delta = (b * h_k, Q_NUM_BLKS, K_NUM_BLKS)
-        _compute_delta_s_kernel[grid_delta](
-            q_mean,
-            K_rot,
-            delta_s,
-            q_mean.stride(0),
-            q_mean.stride(1),
-            q_mean.stride(2),
-            q_mean.stride(3),
-            stride_kb,
-            stride_kh,
-            stride_kn,
-            stride_kd,
-            delta_s.stride(0),
-            delta_s.stride(1),
-            delta_s.stride(2),
-            delta_s.stride(3),
-            h_k,
-            s_k,
-            d,
-            BLOCK_N=BLOCK_SIZE_M,
-        )
-
-    return Q_rot, K_rot, delta_s
